requirements.txt
scrapy==2.6.2
flask==2.0.3
PyDispatcher==2.0.6